% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGL2-package.R
\docType{package}
\name{SGL2-package}
\alias{SGL2}
\alias{SGL2-package}
\title{SGL2: Fit a GLM (or Cox Model) with a Combination of Lasso and Group Lasso Regularization}
\description{
Fit a regularized generalized linear model via penalized maximum likelihood. The model is fit for a path of values of the penalty parameter. Fits linear, logistic and Cox models.
}
\author{
\strong{Maintainer}: Yuxi Yang \email{15364051195@163.com}

Authors:
\itemize{
  \item Yuxi Yang \email{15364051195@163.com}
  \item Noah Simon
  \item Jerome Friedman
  \item Trevor Hastie
  \item Rob Tibshirani
}

}
\keyword{internal}
