% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_linear_fit.R
\name{sgl_linear_fit}
\alias{sgl_linear_fit}
\title{Fit a linear sparse-group lasso model.}
\usage{
sgl_linear_fit(
  X,
  y,
  group_index,
  lambda,
  alpha = 1,
  group_weight = NULL,
  step_size = NULL,
  max_iter = 1000L,
  tol = 1e-08,
  fit_intercept = TRUE,
  initial_beta = NULL,
  initial_intercept = NULL
)
}
\description{
The optimized objective is:
}
\details{
1 / (2n) * ||y - intercept - X \%*\% beta||^2 +
lambda * (
alpha * ||beta||_1 +
(1 - alpha) * sum_g w_g ||beta_g||_2
)
}
