% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_logistic_fit.R
\name{sgl_logistic_fit}
\alias{sgl_logistic_fit}
\title{Fit a Logistic sparse-group lasso model.}
\usage{
sgl_logistic_fit(
  X,
  y,
  group_index,
  lambda,
  alpha = 1,
  group_weight = NULL,
  step_size = NULL,
  max_iter = 1000L,
  tol = 1e-08,
  fit_intercept = TRUE,
  initial_beta = NULL,
  initial_intercept = NULL
)
}
\description{
Fit a Logistic sparse-group lasso model.
}
