% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGLmain.R
\name{SGL}
\alias{SGL}
\title{Fit a sparse-group lasso model.}
\usage{
SGL(
  data,
  index,
  type = "linear",
  maxit = 1000,
  thresh = 0.001,
  min.frac = 0.1,
  nlam = 20,
  gamma = 0.8,
  standardize = TRUE,
  verbose = FALSE,
  step = 1,
  reset = 10,
  alpha = 0.95,
  lambdas = NULL
)
}
\arguments{
\item{data}{A list containing the data.}

\item{index}{A numeric vector containing the group labels.}

\item{type}{The type of model to fit.}

\item{maxit}{The maximum number of iterations.}

\item{thresh}{The convergence threshold.}

\item{min.frac}{The minimum fraction of non-zero coefficients.}

\item{nlam}{The number of lambda values to try.}

\item{gamma}{The gamma value.}

\item{standardize}{Whether to standardize the data.}

\item{verbose}{Whether to print progress messages.}

\item{step}{The step size.}

\item{reset}{The number of iterations between resets.}

\item{alpha}{The alpha value.}

\item{lambdas}{The lambda values to try.}
}
\description{
Fit a sparse-group lasso model.
}
